Rename things named `Ajax` to `AdminAjax`
The original naive implementation of this plugin assumed that only one
class would be relying on AJAX. Working under that assumption, I had
named files, classes, variables, etc. in a non-specific way. For
example, the AJAX requests were being made by a class called `Ajax`.

This practice has caused a tight coupling between the `Ajax` class and
the `AdminUI` class. This is a problem because the next feature that
will be built is an AJAX filter module that needs to make AJAX requests
from the public side of the app.

The solution to this problem that I've chosen to go with is another
fairly naive approach. I decided to simply rename anything that had
"AJAX" in the name. For example, the constant that was originally named
`$AJAX_JS_ID` has become `$ADMIN_AJAX_JS_ID`. By appending the `ADMIN`
prefix, the existing code will at least look differently than the new
AJAX code that will be coming in a subsequent commit.

This approach has some obvious drawbacks, *namely*, much of the AJAX
logic will be duplicated after the filter is implemented on the front
end. Depending on if/how the requirements for the filter change in the
future, it may be worthwhile to extract the lower-level aspects of the
AJAX requests into a new object, and share that object around to
whichever classes need to make AJAX calls.

Another drawback is that the renaming strategy doesn't actually do
anything to decouple the `AdminAjax` (formerly just `Ajax`) and the
`AdminUI` classes. One possible approach might be to extract a `Post`
object that is composed of the `AdminAjax` and `AdminUI` classes. The
`Post` object might even be a more appropriate place for the AJAX code
that updates the "Edit Brochure" page.