Document enhancement to filtering behavior
Preliminary documentation on something that should be addressed in the
future. Without any specific insight into why this bug is occurring, I
chose to compile a descriptive "TODO" comment, rather than creating a
full-blown issue.

Once someone has a chance to investigate this a little deeper,
extracting the TODO and whatever supporting information exists into an
issue would be a great idea.