Fix URL string after AJAX requests
This enhancement rewrites the current URL after filtering brochures.

A user might land on the brochure page and click through several pages
of brochures. For example, on page 3, the URL might look like this:

  `https://www.clearymillwork.com/resources/brochures/pages/3`

If at that point, the user decides to use a filter, an AJAX request is
made. Once the data is returned, the page contains the first page of
brochures - but the URL still appears to indicate page 3.

Now, the URL of the base page is returned along with the brochure data,
and the browser state is updated to reflect the page content.