Add checkout sequence to brochures page
This commit introduces a feature for adding brochures to a "cart", and
then eventually requesting via direct mail. This enhancement depends on
a small JavaScript library for interacting with cookies. The decision to
use a dependency here was more about building this feature quickly than
it was about needing a robust solution. If the need for better cookie
APIs comes up in the future, consider doing away with this library.